NO DESCRIPTION

+++ src/Libraries/CTM.Services/StatisticsReport/TradeTypeProfitEntity.cs
@@ -12,16 +12,16 @@ namespace CTM.Services.StatisticsReport
         public int TradeType { get; set; }
         public string TradeTypeName { get; set; }
         public DateTime TradeDate { get; set; }
-        public double CurValue { get; set; }
-        public double MondayValue { get; set; }
-        public double DayFund { get; set; }
-        public double DayProfit { get; set; }
-        public double DayRate { get; set; }
-        public double YearAvgFund { get; set; }
-        public double YearProfit { get; set; }
-        public double YearRate { get; set; }
-        public double AccAvgFund { get; set; }
-        public double AccProfit { get; set; }
-        public double AccRate { get; set; }
+        public decimal CurValue { get; set; }
+        public decimal MondayValue { get; set; }
+        public decimal DayFund { get; set; }
+        public decimal DayProfit { get; set; }
+        public decimal DayRate { get; set; }
+        public decimal YearAvgFund { get; set; }
+        public decimal YearProfit { get; set; }
+        public decimal YearRate { get; set; }
+        public decimal AccAvgFund { get; set; }
+        public decimal AccProfit { get; set; }
+        public decimal AccRate { get; set; }
     }
 }
